// src/ts/model/disciplinesDefinitions.ts
/** Kai series disciplines codes */
export enum KaiDiscipline {
    Camouflage = "camflage",
    Hunting = "hunting",
    SixthSense = "sixthsns",
    Tracking = "tracking",
    Healing = "healing",
    Weaponskill = "wepnskll",
    Mindshield = "mindshld",
    Mindblast = "mndblst",
    AnimalKinship = "anmlknsp",
    MindOverMatter = "mindomtr"
}

/** Magnakai series disciplines codes */
export enum MgnDiscipline {
    Weaponmastery = "wpnmstry", // Duplicate!
    AnimalControl = "anmlctrl",
    Curing = "curing",
    Invisibility = "invsblty",
    Huntmastery = "hntmstry", // Duplicate!
    Pathsmanship = "pthmnshp", // Duplicate!
    PsiSurge = "psisurge",
    PsiScreen = "psiscrn",
    Nexus = "nexus", // Duplicate!
    Divination = "dvnation"
}

/** Grandmaster series disciplines codes */
export enum GndDiscipline {
    GrandWeaponmastery = "wpnmstry", // Duplicate!
    AnimalMastery = "anmlmstr",
    Deliverance = "deliver",
    Assimilance = "assimila",
    GrandHuntmastery = "hntmstry", // Duplicate!
    GrandPathsmanship = "pthmnshp", // Duplicate!
    KaiSurge = "kaisurge",
    KaiScreen = "kaiscrn",
    GrandNexus = "nexus", // Duplicate!
    Telegnosis = "gnosis",
    MagiMagic = "magi",
    KaiAlchemy = "alchemy"
}